started the float numbers

+++ inc/nap_types.h
@@ -24,16 +24,16 @@
  * The corresponding C type of a nap-script "int" type. This is a 64 bit signed 
  * integer.
  */
-typedef int64_t  nap_int_t;
+typedef int64_t     nap_int_t;
 
 /* the NAP type of a real number */
-typedef double   nap_real_t;
+typedef long double nap_real_t;
 
 /* the NAP type of a string */
-typedef char*    nap_string_t;
+typedef char*       nap_string_t;
 
 /* the NAP type of a byte */
-typedef uint8_t  nap_byte_t;
+typedef uint8_t     nap_byte_t;
 
 /** @} */
